Default to first tab if active tab not available
Summary:
This is only really an issue in theUIDebugger where we have a dynamic number of tabs. on some nodes you have 3 tabs, on others you have 2. if you select the 3rd tab and move to a node with 2 tabs nothing will be selected.

With this fix we will fallback to the first tab

Reviewed By: lblasa

Differential Revision: D50595977

fbshipit-source-id: 25c36a5609f774c2d3fd650eb6520104864fbe95