default adb path
Summary:
As I was setting up a new mac I had to change this setting in flipper settings.

From a few android people that I've spoken to no one could answer me if the old one is still used.

Ideally we should infer this during the first time flipper starts

- `/opt/android_sdk` current default
- `$HOME/Library/Android/sdk` adb from Android Studio
- `$HOME/fbsource/third-party/toolchains/android-sdk` adb that is always available in fbsource

Reviewed By: lblasa

Differential Revision: D51120929

fbshipit-source-id: edb2a58b9c9f37465ea2fc5493975dd427d5523b