Simplify shutdown logic
Summary:
The previous logic aimed to reuse an existing server during bootstrap if the launched version was higher than the running one. This is no longer required or wanted.

## Risk assessment: LOW
### Rationale
It is extremely rare to launch Flipper whilst already having another instance running. This can happen during development, but it is extremely rare in production.

Launcher (singleton) launches Server (singleton).
Launcher can be executed multiple times and this will not create newer server instances.

If anything, if we are unable to kill any other instance, whatever that may be, continue. This is to cover the cases where a shutdown may have been acknowledged but the process is still shutting down.

Reviewed By: antonk52

Differential Revision: D51232901

fbshipit-source-id: 8b8b85f4bac68f5670b1878e827871f78dc68999