Ask to restart flipper to pick up ios SDK
Summary: Running a check is run it only executes the `run` function. The `run` function for `ios.sdk` is a pure function that detects if ios SDK available from `EnvironmentInfo`. This info is picked up when flipper first started. Even if you installed ios SDK flipper won't pick it up while running. Restarting flipper will pick up the SDK and make it available. We should be clear about it.

Reviewed By: mweststrate

Differential Revision: D51614258

fbshipit-source-id: 8b2a8b5b94ec0038acf9f383e55e166c52b2ccab