Only enable kb controls for tree when tree is focused
Summary:
Previously kb controls where always listening, this means if you were to press left arrow in an input of the sidebar it would collapse a tree node.

Now make the tree focusable, it automatically focused when any child is clicked, but to retain the previous behaviour of click the visualizer and then use kb to move around we focus the div in an effect.

Changelog: UIDebugger: Add live editting of attribute on the running app from the side bar. Currently Native android only

Reviewed By: lblasa

Differential Revision: D51623365

fbshipit-source-id: f6a2fbe0be264623dc7a5a508be9a6181cdf6c04