Re-request env options on every doctor run
Summary: Currently installing new ios and android SDKs require users to restart flipper to pick up these changes. In reality the restart is not required. As we currently cache the initial env info for the entirety of flipper server lifecycle. Since flipper doctor is only run once during the launch and can be explicitly triggerend by a user it is better to gather env info on every doctor run instead of requesting user to restart flipper.

Reviewed By: passy

Differential Revision: D51808597

fbshipit-source-id: da488a36a9fbff4c084378647099048fa0c2b767