Fix MacOS signing issues
Summary:
Context in https://fb.workplace.com/groups/1100215790030899/posts/7208076185911465

> ...peertalkMac.framework: bundle format is ambiguous (could be app or framework)

I initially fixed the symlinks and confirmed it worked (P897979464), ~~but then when I ran it in Sandcastle, signing failed. After debugging it, I noticed the build-script resolves the symlinks (aka, does not honor them and copy the target files/dirs). Because `PortForwardingMacApp` frameworks are quite basic, and they do not need to use the `Versions` schema, lets just remove it.~~

**UPDATE**: fixed the build script (keep symlinks); we can now move forward with the `Versions` fix.

> ...Flipper.app: code object is not signed at all
> In subcomponent: .../Flipper.app/Contents/MacOS/FrameworkVersion.txt

You cannot have non-signabe artifacts there; it used to be just a "bad practice" but now codesign enforce it as a failure.

I didn't find any reference to it (quick `zbgs`), so it should be safe to move it to the `Resources` folder.

Reviewed By: lblasa

Differential Revision: D51887279

fbshipit-source-id: 4ebe0885cfaa50ccb958de7bf5075eeb7e0e9b98