Report connecting apps from the TLS socket
Summary:
This was missing. Only apps going through the certificate exchange process were marked as 'apps trying to connect'.

This changes fixes it.

Notes:

In my tests it is not always visible as the app can establish the connection very quickly. So you may be wondering what the point is:
- Consistency across our connection process.
- It will become useful of there's a failure coming from an app that already has the client certificate.

Reviewed By: Katalune

Differential Revision: D51984388

fbshipit-source-id: 44796ed5666f4570adbfcf455063c5a56936778f