DEBUG_LOG macro removed
Summary:
DEBUG_LOG was a facility in which, if FLIPPER_DEBUG_LOG was defined, it would have printed things into the console.

Instead, it is convenient to capture all those logs, even if not printed, as to be sent to Flipper Desktop.

As such, define a 'log_debug' function which will always collect the logs for debugging.

If FLIPPER_DEBUG_LOG is defined, then it will also log it to the console.

Note: as a reminder, for the reviewer, the reason this is needed is because Flipper logs can become VERY noisy in the case a connection is not established with Flipper Desktop. This is the majority of cases where engineers are debugging their apps without Flipper.

Reviewed By: passy

Differential Revision: D52017948

fbshipit-source-id: d8cc08b505ab29974335b6862e76854adac2a158