Send client logs
Summary:
Until now, app (client) logs stayed within the app (client). Only visible when examining device logs (if at all).

This change changes this, at least in the easiest way which can yield the most impact.

- Collect the logs coming through logger into a container.
- Whenever to try to exchange certificates, send the collected logs until that point.

The consequence of this is:
- Any state and error logged before exchanging certificates will be sent to Flipper.
- After this point, any state or error which prevents establishing the TLS connection will get collected.
- Once the limit of failed connection attempts is reached, the client will attempt to exchange certificates again in which case, all these logs preventing it to establish a TLS connection will be sent.

Note about circular buffer:

Use a very simple approach to circular buffer. The app will constantly try to connect even if Flipper is not running, which may represent the vast majority of cases. If we were to keep all logs, we would soon run out-of-memory as the reconnect cycle takes place after a couple of seconds. Also, don't truncate the container once capacity is reached because this could truncate valuable logs if the app ends up trying to connect to Flipper.

Reviewed By: passy

Differential Revision: D52017677

fbshipit-source-id: 941b184bf1490ae8d8711bdae23671d41551c6df