Sign Certificate Ack
Summary:
Until now, once the certificate signing request response was sent to the app, Flipper was kept in the dark and assumed the client was good to establish a TLS connection. That assumption may not always be true.

This diff adds a new method called "signCertificateAck" to the server-types.tsx file in the Flipper common library.

This message is used to acknowledge the receipt of a certificate signing response from the Flipper desktop application by the client. It gets sent just before closing the connection.

Reviewed By: passy

Differential Revision: D52017722

fbshipit-source-id: 0207eaec4e5f1dabb5b63a5f9aebcd1fe9bb4c4c