close flipper ui on platform change save
Summary:
Since the cocoa app does not support restarting flipper. Multiple places including "platform selection" calls to to restart do not actually restart the app, but kills the server. This causes confusion for the users as it leaves the UI running without the server with "cannot connect to server". As a temporary workaround untill we can restart flipper from the cocoa app, let's kill flipper UI as well to force user to restart flipper.

Motivated by the support post
https://fb.workplace.com/groups/flippersupport/posts/1743869656093685

Reviewed By: LukeDefeo

Differential Revision: D52202802

fbshipit-source-id: 3585da2d04a67c901df8630255e694a10a1b699c