Strip port forwarder from flipper-server npm bundle
Summary:
Our npm releases have been failing for a while now: https://github.com/facebook/flipper/actions/runs/7209384441/job/19640603242

This is because the bundle we try to upload contains various symlinks:

```
❯ find . -type l
./static/PortForwardingMacApp.app/Contents/Frameworks/peertalkMac.framework/Versions/Current
./static/PortForwardingMacApp.app/Contents/Frameworks/CocoaAsyncSocketMac.framework/Versions/Current
./static/PortForwardingMacApp.app/Contents/Frameworks/FBPortForwarding-Mac.framework/Versions/Current
```

I can't think of an easy way to fix this as MacOS expects these to be symlinks and npm will block us if we have them. In the medium-term, we probably need to bundle the app in an extra zip or tarball and extract it on demand somewhere.

For now, this will simply remove it and break a bit of iOS functionality which seems better than having everything broken.

Reviewed By: lawrencelomax

Differential Revision: D52253677

fbshipit-source-id: 61c09799a9d4cefc1c02f163ac1efa883566a8c5