Fixed plugin installation table columns width (#5161)
Summary:
the issue was that in the plugin installation popup table, the columns had a width of 0, making it impossible to see any data and the install/uninstall button.

This PR addresses the issue https://github.com/facebook/flipper/issues/5124

<img width="1379" alt="Screenshot 2023-09-21 alle 16 46 19" src="https://github.com/facebook/flipper/assets/29179493/cf14a3b2-7385-4677-ade8-9e23108ca9ec">

After the fix

<img width="1379" alt="image" src="https://github.com/facebook/flipper/assets/29179493/08caacf4-8326-429e-bd9c-a50feb55cea6">

I resolved the issue by moving the constants declaring the column widths inside the component, which previously caused the columns to have a width of 0 in the plugin installation popup table.

## Changelog

Pull Request resolved: https://github.com/facebook/flipper/pull/5161

Reviewed By: antonk52

Differential Revision: D52599827

Pulled By: passy

fbshipit-source-id: 175163b5e2b323c7dd476c7ffb8e97e94be5ee17