Fix socket disconnect during certificate exchange
Summary:
Before this change, the socket will not attempt to reconnect if there was an error during the TLS connection.
The problem is that a socket disconnect will not call the event handler and it is the event handler that sets all the correct state.

This change manually invokes the socket event handler, the alternative was to also change the state after the disconnect, but that just spreads the state logic into different places which makes the code harder to reason about.

Reviewed By: aigoncharov

Differential Revision: D52217852

fbshipit-source-id: 5c96631045fa187d49b3316746c7c5ea1fbaefda