Convert xplat/sonar/android/sample/src/debug/java/com/facebook/flipper/sample/FlipperInitializer to Kotlin
Summary:
This was an automatic conversion. We need this to be converted to Kotlin to unblock D52563388 , given we are stuck in the current litho OSS release.

By converting this class to Kotlin, I can use `ComponentsConfiguration.isDebugModeEnabled` because the public API is "intact". If I had to call it from Java, `Flipper` would complain because the calling method is actually `setDebugModeEnabled` which corresponds to a breakage of the current API.

Reviewed By: astreet

Differential Revision: D52658356

fbshipit-source-id: 7e5bd03fd2ca1e2f20468d25637bbe5f603be654