Fix emission of incremental mount debug event
Summary:
hannahyang pointed me to a situation where we could see IncrementalMount related events on the UIDebugger even with Incremental Mount disabled. After a bit of debugging I realized I had added the events in a not correct place and they should be included at a lower level.

In this diff I fix the events location.

Differential Revision: D52871459

fbshipit-source-id: b2643f1ab507e27d5a3dda7839634c447a64fb88