Workaround for loading native library crash
Summary:
Supporting recomposition counts in Flipper requires adding a native library (`libart_tooling.so`) which is a part of `"//third-party/java/androidx/inspection/inspection:inspection"`. It seems like for some reason this library fails to be loaded on some devices: https://fb.workplace.com/groups/instadroiddiscussions/posts/7838944066120737

I've checked and the .so files for all architectures are present:
https://pxl.cl/4dcr8

so at this point it's unclear why it sometimes fails to load.

Prior to this change the library was eagerly loaded at the moment when `ComposeViewDescriptor` is loaded.

This diff addresses the crash by wrapping all code that ends up loading native libraries in try-catch blocks and it also delays loading `libart_tooling.so` until Flipper calls `ComposeViewDescriptor.onGetChildren` for the first time.

Reviewed By: passy

Differential Revision: D53042369

fbshipit-source-id: 8f4a1d11a8d8acba0db1810d723fc1b7ac07898d