Fix enabling composition info
Summary:
Compose 1.6 changed how the slot table is enabled on ComposeView. The slot table is required for the tooling to be able to dump compose node hierarchy. In D53085052 I've implemented the new way of enabling this information, but back then we were still using Compose 1.5.x at Meta so I wasn't able to properly test it. Now Compose 1.6 landed and it turns out that the new logic is executed too late - it is enabling slot table but the compose node tree already exists so the table is empty when we open the compose screen for the first time (when the screen is closed and opened again the issue is fixed, which confirms that this is the issue with an empty slot table).

Fix - in this diff we're eagerly enabling slot table on ComposeView at the moment when Flipper Compose Plugin is registered which usually happens during app startup.

Reviewed By: pengj

Differential Revision: D53123294

fbshipit-source-id: 0f7f33379abaeb7337c41a761267d6898f9f3688