Add support for streaming responses to Flipper
Summary:
We discovered that apps in debug mode would not incrementally render streamed responses.
This was caused by the flipper interceptor reading the full response before passing it to back to the app.

This diff copies the response as it is read by the application, and sends it to flipper after the response is closed by the app logic.

If the app does not read the response, the result is not sent to flipper.

Reviewed By: gorodscy

Differential Revision: D53150417

fbshipit-source-id: 9c78d1045adeedff8d5d0b1a59269ee24f621685