Handle multiple plugin urls
Summary:
Refactor the marketplace plugin download interface to allow for multiple URLs to be specified. They will be tried sequentially and allows for multiple mirrors to be used. Right now, only one URL is provided for each.

Changelog: Plugin marketplace downloads now expect a  `downloadUrls: string[]` instead of a singular url for fallbacks

Reviewed By: lblasa

Differential Revision: D53223203

fbshipit-source-id: 0d9c8ae4c3ab04f79ca45372f0ae9fbbcb187291