Fix visualiser sizing quirk
Summary:
Previously the visualser got its size from its parent. In the main case this was provide by the flipper _Sidebar component which can be resized.

The issue is the onResize for this was more of a user suggestion than the actual size. This could result in a situation where the visualiser thought it had more width than the browser actually gave it. This lead to it overflowing in the horizontal axis which we dont want to happen

Solution is no longer provide the visualiser the width it should use. Instead it measures its width and reacts to changes. This fixes the issue above and also means that when the visualiser is placed in the right side bar (for the framework events scenario) it reacts correctly to that resizing

Reviewed By: elboman

Differential Revision: D53355309

fbshipit-source-id: 767a9283198c0037746d276621aecf99e222862a