Keep showing frameowrk events timeline when node goes offscreen
Summary:
Previously when a node goes offscreen we lose all the associated data with it, the frameowrk events are still there but we needed some bits of the node to display the tab.

Now we store the entire node in the selection rather than the ID. this allows us to continue to show the framework events tab. Usefull for debuging visibility / unmount behaviour

We deliberated check and dont show the node attributes when its offscreen / not in the frame. This is because this data is no longer valid and we would be instead showing the last known node state which would be misleading

Changelog:  UIDebugger - Keep showing node level event timeline when node goes offscreen

Reviewed By: elboman

Differential Revision: D53352198

fbshipit-source-id: 48fbd1515b9a75083f0987ee21cd5f92ebd0c3cd