Add horizontal scroll bar for visualiser
Summary:
A few users complained that the visualiser was cut off when it was enlarged past a certain point. While you could shrink it to fit ( we preserve aspect ratio) this isnt that obvious.

To fix i have put the visualiser in a scroll container. The scrollbars interfere with the visualiser so i added some horizontal padding

Changelog: UIDebugger add vertical scroll bar to visualiser when height exceeds viewport

Reviewed By: mweststrate

Differential Revision: D53308772

fbshipit-source-id: f44f0a3a0aa62ffbf2d93fc0e22e5e85cb789c80