Add RecyclerBinderConfig in the Attributes Tab
Summary:
We want to provide a quick way to see the `RecyclerBinderConfig` associated with a given `Recycler` (which is the abstraction used in Sections/LazyList).

I'm making use of reflection (uggggh I know .. but right now it is the only way) to access the `RecyclerBinderConfig` and display it in the attributes tab associated with a `Recycler`.

Reviewed By: adityasharat

Differential Revision: D53308774

fbshipit-source-id: fc6d1a776365eda7a773ca192845c85267dca20f