Workaround for a crash when converting Compose nodes
Summary:
For an unknown reason we're getting reports about `NoSuchMethodError`s happening when using ui-inspector code. I'm not able to reproduce it locally so I'm adding another try-catch to prevent it from crashing.
I hope that once we convince Google to properly publish ui-inspector as a standalone library and we won't have to copy this code manually to our monorepo, those issues will get resolved.

This diff also updates the ui-inspector code in flipper codebase which is needed for gradle builds.

Reviewed By: pengj

Differential Revision: D53701915

fbshipit-source-id: aced2a1ca311c6d5db01e48903ef9662649811b4