Reword "Server Exchanged"
Summary:
In some cases, Flipper is unable to match the connection of an incoming application to a specific connected device.

This is guaranteed to happen when WWW certificate exchange is used. In this case, the device title was a combination of the app name with "Server Exchanged". This is an implementation detail which can only confuse our users.

Instead, just use "Unknown Device" as this is at least more clear and not tied to WWW exchange. This will also be used in cases in which we managed to connect to an application but we are unable to match the device even when using FS exchange. Not used on this diff, but it will later on in the stack.

Reviewed By: elboman

Differential Revision: D53849410

fbshipit-source-id: 58c00491fb9e14f79b74e4fdf9d6aa89de68c9b1