Show bloks state in sidebar
Summary:
context D53574560

We join the variable id on the client to variable id in redution trace.

The interesting part is when there is a scope key

This happens when there are dynamic children on the client, this can be due to async components substituting loading screen for content, tree mutations children bindings etc

Since all state in bloks is stored at the tree manager level, any variabled declared inside of scoped (dynamic children) need to be scoped with their key path. This is because the server can't know ahead of time how many dynamic children there will be and therefore how many ids to generate

The client suffixes the scope key to the server variable id so we must do the same

Reviewed By: antonk52

Differential Revision: D53665302

fbshipit-source-id: 366a5f357a5e950ff044dddddff71d4673af3893