Visualiser overlay refactor
Summary:
The visualiser code was gettign out of hand >900 lines.

There was a common theme of various overlays on top of the wire frameframes these are extracted to a component. this make it easier to see the visulaiser is made of 3 layers

Snapshot
Wireframes
Overlays

VisualisatioNode2D renamed to VisualisationWireframeNode for clarity

Reviewed By: elboman

Differential Revision: D53813068

fbshipit-source-id: 23864b03cd2715325b70c9311294bf7a157c6945