Fix issue with trailing overlays
Summary:
We were using the node selection to drive these overlays, node selection is allowed to go stale w.r.t to teh current frame so we can carry on displaying the framework events sidebar. This results in misplaced overlays when the selected node goes offscreen.

To fix we resolve the node selection against the current frame

Reviewed By: elboman

Differential Revision: D53813069

fbshipit-source-id: db030c29798744ebeb9ac51af6d7824bdabbc862