enforce template string to +
Summary:
over the last year we had a multiple breakages small or large due to misuse of overloaded `+` which can result in expected but unintuitive behaviour for a developer. This eslint rule should remove this confusion.

For a reviewer see .eslintrc.js, the rest changes were made by `yarn run lint:eslint --fix`

Reviewed By: lblasa

Differential Revision: D54760654

fbshipit-source-id: 83f18374917994bcb70a9aeed851057a15061d34