ensure E2EDumpsysHelper can work with nested Litho trees
Summary:
Nested trees use `LithoRenderTreeView` (a subclass of `BaseMountingView`) to mount the layout state. However, our dumpsys helpers (which are used by the E2E framework and Flipper) make assumptions that the mounting view is always a `LithoView`. Unfortunately, this does not work.

The nested tree experiment ended up breaking this assumption as well as a lot of tests that suddenly couldn't dump the view/component hierarchy correctly

This diff changes the dumpsys helper(s) to use `BaseMountingView` instead which covers all mounting views used in the Litho framework

Reviewed By: adityasharat

Differential Revision: D54789839

fbshipit-source-id: ef18bd4b87cdd7a5cee58f9720438bf521b2a4ee