Retry after a delay
Summary:
If there's an error, retry after a one second delay as to smooth out the user experience i.e. if the QR code is invalid, decryption will fail very quickly and this will result in just a blink on the the "Reading" state.

With this change, error is shown for a second which is enough and provides a better UX.

Reviewed By: antonk52

Differential Revision: D55015643

fbshipit-source-id: 3e2d99c644178d24613d18c3c13648a205366515