Enable QR exchange, if possible
Summary:
QR exchange is only possible if the encrypted certificates are being sent on the response and the client is iOS.

GK: https://www.internalfb.com/intern/gatekeeper/projects/flipper_encrypted_exchange/

If the conditions are met, then use the backup certificate provider which is only set on iOS too.

# How does this work?

The base certificate provider will always generate a bundle containing client and server certificates that need to be transferred back to the client for connection.

If Flipper succeeds to transfer these files, then no error takes place, and nothing else is required.

However, if the certificate provider throws an error during exchange **AND** the user is on the mentioned GK **AND** the connecting app is an iOS app, only then Flipper will send an encrypted bundle with the required certificates.

Once the client receives the response, it will check for errors and use any set certificate provider to complete the exchange, as usual.

If, after this process ends, the client doesn't have the necessary certificates **AND** the encrypted certificates bundle is part of the response **AND** a backup certificate provider is set (QR certificate provider), only then it will try to use said provider to decrypt the encrypted certificate bundle.

## QR Certificate Provider.

1. Shows an alert to the user indicating the app is trying to connect to Flipper.
2. Connection can be dismissed for the session.
3. If the user decides to connect to Flipper, the camera viewer will be shown and the QR code can be scanned from Flipper Desktop.
4. If the QR code contains a valid decryption key, then the encrypted certificate bundle is decrypted and the exchange completes.
5. If the QR code is invalid then the user can try again until this is read correctly.
6. The user can dismiss the camera at any time.

### Note
If the encrypted bundle is not sent back to the client, then the client keeps behaving as always. Therefore, it is only necessary to switch off the GK for the encrypted bundle not being sent to the client so this is the feature kill flag.

# Rollout proposal

1. Flipper team to test it and just ensure nothing out of the ordinary is observed.
2. Ask a few engineers which have failed to use WWW exchange to give this a test i.e. add them to GK and instruct them to test.
3. Bloks engineers partial rollout, GK maybe at 50%.
4. Increase rollout, partially, until 100% is reached.
5. After a cooldown period, remove WWW exchange altogether from our codebase i.e. clients and desktop. This includes:
 5.1. Remove WWW toggle from Internal Settings
 5.2. Remove Desktop WWW certificate provider
 5.3. Remove iOS WWW certificate provider
 5.4. Remove WWW endpoints for both FB and IG

Reviewed By: antonk52

Differential Revision: D54901034

fbshipit-source-id: 51eebdfe71847cafa9a6482fa62859cced038aab