Limit the impact of the connection notifications
Summary:
These notifications can get a bit annoying and in the way.

They are now keyyed on app so multiple notifications from the same app do not pile up

They are now allowed to cover the nav bar. Better to cover this than cover the plugin content area

Added a limit also

Reviewed By: mweststrate

Differential Revision: D55250950

fbshipit-source-id: c04d2efc88fad4d4c9b7f8278d85c2e7facedb35