Use String as Id instead of integer
Summary:
Using integers as Id was not a good idea. This is because the size of int on iOS is 64bits is 64bits unsigned long, where as the max safe int in a js is 9007199254740991

What happens whne you send  a large int than this is that the number is truncated.

Coupled to this NSString hash function is not randomly distributed and share a common prefix, this causes  us to get collisions in ids in bloks where share a common client id prefix.

To fix we send ids as strings, this allows us to represent numbers of any length at a small performance penalty

Reviewed By: lblasa

Differential Revision: D55692665

fbshipit-source-id: 2fc0725fbd14adb681a328ca9f5f119a7d088199