Align tree nodes with dummy vscode icon
Summary:
An overlooked side effect of moving the vscode icon to the left of the text means that for nodes that dont have position info they appear idented 1 level less than their siblings due to the lack of an icon which is about the size of the tree indent.

This solution isnt ideal but it means that the siblings line up properly.

Differential Revision: D56066784

fbshipit-source-id: 29410914ce984d8885de335d7bb468dea5de5d85