Turn on auto scroll and force sort by timestamp ascending
Summary:
Sometimes events arrive out of order from the client

Today the timestamp column is sortable but by detault out of the box no sorting is applied, the user has to know to set it one way or another.

This can lead to events appearing in the 'wrong' order  easily

This diff makes some decisions for the user, we enforce ascending sort on the timestamp column and turn on the auto scroller by detault which is a great feature to keep the most recent content on screen.

Overall this is consistent and better experience

Reviewed By: mweststrate

Differential Revision: D56186052

fbshipit-source-id: 5b645ff99518cd455a6c3d29befbe7f5750f1ef6