Find device by serial for incoming connection instead
Summary:
This is a better approach to create dummy devices for connections with a device serial which doesn't match any of the available connected devices.

This approach was reserved for WWW exchange, but it is better to generalise for all type of exchanges.

If the app is connected, is providing the device identifier used during the certificate exchange. If is connecting but the device does not appear to be available, create a dummy one instead.

This is better than later on rejecting the connection.

Reviewed By: LukeDefeo

Differential Revision: D55797637

fbshipit-source-id: 78ec6fa6f2bfdb0b0628d814594cba51aa6b3074