Restart ios app automatically after selecting a sandbox
Summary: Android apps do not need to be restarting app when changing sandbox. Currently ios app gets killed when we set a new sandbox. This requires user to switch back to the device and start the app manually which is unnecessary friction. This diff exposes `runDeviceAction` on flipperLib to be able to launch the app directly from a plugin.

Reviewed By: LukeDefeo, lblasa

Differential Revision: D56548543

fbshipit-source-id: 10c769d6114a278d10a6174aa9f2b57cd5636558