Prevent UILabel from becoming active child
Summary:
context S415547

On ig there is a random UILabel that can beceome the last child of the ui window.

This can become active and therefore we dont traverse the correct view.

This attempts to mitigate this by making sure we only consider a view for becoming an active child if it itself has subviews

Reviewed By: dinhvh, lblasa

Differential Revision: D57058208

fbshipit-source-id: 57fcbc3cc6ed2ba02811d73bd5effb8db0b88bd7