Remove usage of ultralight from NT uidebugger
Summary:
using Ultralight is problematic for uidebugger since UIDebugger is used in contexts where ultralight isnt (Open source & flipper / litho sample apps in buck)

The new approach is a tad hacky, we add a descriptor that handles debug component (the litho wrapper for litho Component class), this new descriptor can also handle NT if the underlying component is an NT component, this is recognised by the WrappedNTComponent class.

If so we add the additional parts for the desktop. This descriptor is manually added in apps that have NT, currently just fb4a

The hacky part is we overrite the original descriptor in the registry, but given how DebugComponent works and is used regardless of if we are NT / Litho i dont see a better way

Differential Revision: D57105063

fbshipit-source-id: 63e291c22c566ff52277013fe1b945f8c41464c1