Replace ScatterSet with a LinkedHashSet
Summary:
We don't expect to have a lot of elements in this set so using ScatterSet wouldn't make a difference here. And it turns out that OSS version doesn't have a proper dependecy added so using ScatterSet breaks OSS CI build.
In this diff I'm replacing it with a regular LinkedHashSet from stdlib.

Reviewed By: pengj

Differential Revision: D57197303

fbshipit-source-id: e7e9dc9dcab07fcb26b9f26e7fe7a78595ec63a5