Perform reflective parameter inspection for additional inspection nodes
Summary:
This diff uses the new Flipper API that allows for performing additional computation related to attributes on demand.

By default the reflective parameter inspection is turned off which makes Flipper Compose plugin to perform well. User will have an ability to select a specific node in Flipper Desktop UI for which additional computation should be performed. When this happens Flipper Compose plugin will use reflection in order to perform additional analysis only on the selected node.

The reflection is still slow but this way the user is in control and if they need to see additional information about some node, they can opt in.

This diff brings back `ReflectionScope` class and adds back methods that perform reflective analysis to `ParameterFactory`.

Reviewed By: pengj

Differential Revision: D57148571

fbshipit-source-id: e53d734b6f9d6c64f69064104765a8089c4f0203